+++ GRDB Tests/Public/RowModel/MinimalPrimaryKeySingleTests.swift
@@ -130,21 +130,6 @@ class MinimalPrimaryKeySingleTests: RowModelTestCase {
     
     // MARK: - Update
     
-    func testUpdateWithNilPrimaryKeyThrowsInvalidPrimaryKey() {
-        assertNoError {
-            try dbQueue.inDatabase { db in
-                let rowModel = MinimalSingle()
-                XCTAssertTrue(rowModel.UUID == nil)
-                do {
-                    try rowModel.update(db)
-                    XCTFail("Expected RowModelError.InvalidPrimaryKey")
-                } catch RowModelError.InvalidPrimaryKey {
-                    // Expected RowModelError.InvalidPrimaryKey
-                }
-            }
-        }
-    }
-    
     func testUpdateWithNotNilPrimaryKeyThatDoesNotMatchAnyRowThrowsRowModelNotFound() {
         assertNoError {
             try dbQueue.inDatabase { db in
@@ -278,21 +263,6 @@ class MinimalPrimaryKeySingleTests: RowModelTestCase {
     
     // MARK: - Delete
     
-    func testDeleteWithNilPrimaryKeyThrowsInvalidPrimaryKey() {
-        assertNoError {
-            try dbQueue.inDatabase { db in
-                let rowModel = MinimalSingle()
-                XCTAssertTrue(rowModel.UUID == nil)
-                do {
-                    try rowModel.delete(db)
-                    XCTFail("Expected RowModelError.InvalidPrimaryKey")
-                } catch RowModelError.InvalidPrimaryKey {
-                    // Expected RowModelError.InvalidPrimaryKey
-                }
-            }
-        }
-    }
-    
     func testDeleteWithNotNilPrimaryKeyThatDoesNotMatchAnyRowDoesNothing() {
         assertNoError {
             try dbQueue.inDatabase { db in
@@ -331,21 +301,6 @@ class MinimalPrimaryKeySingleTests: RowModelTestCase {
     
     
     // MARK: - Reload
-    
-    func testReloadWithNilPrimaryKeyThrowsInvalidPrimaryKey() {
-        assertNoError {
-            try dbQueue.inDatabase { db in
-                let rowModel = MinimalSingle()
-                XCTAssertTrue(rowModel.UUID == nil)
-                do {
-                    try rowModel.reload(db)
-                    XCTFail("Expected RowModelError.InvalidPrimaryKey")
-                } catch RowModelError.InvalidPrimaryKey {
-                    // Expected RowModelError.InvalidPrimaryKey
-                }
-            }
-        }
-    }
     
     func testReloadWithNotNilPrimaryKeyThatDoesNotMatchAnyRowThrowsRowModelNotFound() {
         assertNoError {